safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A5_T2-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Cannot use the undeclared label 'IN_DO_FUNC'.
S12.7_A5_T2

PASS Expected parsing failure

TEST COMPLETE
